Refatora página Industry para melhorar acessibilidade e performance
