+++ src/app/industry-research/page.module.css
@@ -15,6 +15,18 @@
   text-align: left;
 }
 
+.visuallyHidden {
+  position: absolute !important;
+  width: 1px;
+  height: 1px;
+  margin: -1px;
+  padding: 0;
+  overflow: hidden;
+  clip: rect(0 0 0 0);
+  white-space: nowrap;
+  border: 0;
+}
+
 .title {
   font-size: 2.2rem;
   font-weight: bold;
